rag-bot/pyproject.toml
[project]
name = "rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agents>=1.4.0",
    "fastapi>=0.124.0",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "qdrant-client>=1.16.1",
    "uvicorn>=0.38.0",
]
